Fetch and dispatch analytics data on page level
